requirements.txt
sys
pillow
dearpygui
pytest